finish switch to cl-lib
`cl' is required again for the sake of `lexical-let', but at least for
recent Emacsen only at compile time.  At runtime `cl' is only required
right before the use of `lexical-let', which is on a code path only
used by older Emacsen.

Also add the `cl-' prefix to some symbols.

This completes #43.